Refactor how rocket instance is built
